+++ backend/src/main/java/botobo/core/config/AuthenticationPrincipalConfig.java
@@ -56,7 +56,7 @@ public class AuthenticationPrincipalConfig implements WebMvcConfigurer {
                 .excludePathPatterns("/api/**", PathMethod.OPTIONS)
                 .excludePathPatterns("/api/workbooks", PathMethod.GET)
                 .excludePathPatterns("/api/quizzes/guest", PathMethod.GET)
-                .excludePathPatterns("/api/login", PathMethod.POST)
+                .excludePathPatterns("/api/login/**", PathMethod.POST)
                 .excludePathPatterns("/api/docs/**", PathMethod.ANY)
                 .excludePathPatterns("/api/workbooks/public/**", PathMethod.GET);
     }
